Working on making username accessible across app
