+++ package.json
@@ -9,7 +9,6 @@
   "devDependencies": {
     "autoprefixer": "^10.4.13",
     "postcss": "^8.4.21",
-    "prisma": "^4.12.0",
     "solid-start-node": "^0.2.19",
     "tailwindcss": "^3.2.4",
     "typescript": "^4.9.4",
@@ -18,6 +17,7 @@
   "dependencies": {
     "@auth/core": "^0.2.5",
     "@auth/solid-start": "^0.1.1",
+    "@prisma/client": "^4.12.0",
     "@solidjs/meta": "^0.28.2",
     "@solidjs/router": "^0.8.2",
     "dotenv": "^16.0.3",
